refactor(api): further split module up into separate files
