+++ crates/client/src/lib.rs
@@ -2,7 +2,7 @@ mod http_client;
 mod http_prism_api;
 
 pub use http_client::{PrismHttpClient, PrismHttpClientError};
-pub use prism_api::{PendingTransaction, PrismApi, PrismApiError, builder, types};
+pub use prism_api::{PendingTransaction, PrismApi, PrismApiError, RequestBuilder, types};
 pub use prism_common::{
     account::Account,
     digest::Digest,
